fix: split write_multi out of the required cache store contract
docs/caching.md promises a store implementing only read_multi and write
keeps working, but CacheStoreContract required write_multi too, so an
author following the docs got three NoMethodErrors from the shared test
module. Move the three batching cases into BatchingCacheStoreContract,
included by the three shipped stores; add a minimal two-method store that
proves the required contract passes without it.
